requirements.txt

scikit-learn
seaborn
matplotlib
numpy
Pillow